+++ Assets/Scenes/Build2.0 onboarding.unity
@@ -9006,15 +9006,15 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 1327803687913001269, guid: 76a5fb4b5e7346347ad7df57fbc44430, type: 3}
       propertyPath: m_LocalEulerAnglesHint.x
-      value: 1.000003
+      value: 359.8955
       objectReference: {fileID: 0}
     - target: {fileID: 1327803687913001269, guid: 76a5fb4b5e7346347ad7df57fbc44430, type: 3}
       propertyPath: m_LocalEulerAnglesHint.y
-      value: -0.0004013192
+      value: 163.0402
       objectReference: {fileID: 0}
     - target: {fileID: 1327803687913001269, guid: 76a5fb4b5e7346347ad7df57fbc44430, type: 3}
       propertyPath: m_LocalEulerAnglesHint.z
-      value: 0.00000001121964
+      value: 1.017995
       objectReference: {fileID: 0}
     - target: {fileID: 1409067402764889638, guid: 76a5fb4b5e7346347ad7df57fbc44430, type: 3}
       propertyPath: m_LocalRotation.w
@@ -10745,7 +10745,7 @@ PlayableDirector:
   m_GameObject: {fileID: 1690276446}
   m_Enabled: 1
   serializedVersion: 3
-  m_PlayableAsset: {fileID: 11400000, guid: 6767f379066cd1046a83d817a86d2393, type: 2}
+  m_PlayableAsset: {fileID: 11400000, guid: 7e6cbbca09a049041aee7402b185f345, type: 2}
   m_InitialState: 1
   m_WrapMode: 2
   m_DirectorUpdateMode: 1
@@ -10765,6 +10765,12 @@ PlayableDirector:
     value: {fileID: 1523748355}
   - key: {fileID: -5635341431066554272, guid: 6767f379066cd1046a83d817a86d2393, type: 2}
     value: {fileID: 714231907}
+  - key: {fileID: 6997562848468670863, guid: 7e6cbbca09a049041aee7402b185f345, type: 2}
+    value: {fileID: 714231907}
+  - key: {fileID: 2235190925458089653, guid: 7e6cbbca09a049041aee7402b185f345, type: 2}
+    value: {fileID: 1728441200}
+  - key: {fileID: 8122700741437066153, guid: 7e6cbbca09a049041aee7402b185f345, type: 2}
+    value: {fileID: 2094558843}
   m_ExposedReferences:
     m_References: []
 --- !u!4 &1690276448
