+++ Assets/Timelines/Signals/Pause Timeline.signal.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bbb29a41ece89134b8341c43448c5bcb
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
